requirements.txt
gunicorn
djangorestframework-simplejwt
django_extensions
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-crispy-forms==2.0
django-datetimepicker==3.14
django-filter==23.2
django-js-asset==2.0.0
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
Markdown==3.4.3
Pillow==9.5.0
pipreqs==0.4.13
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
yarg==0.1.9
